Move board specific tests
