+++ apps/test/unit/lib/kits/maker/ButtonTest.js
@@ -2,7 +2,7 @@
 import _ from 'lodash';
 import {expect} from '../../../../util/deprecatedChai';
 import five from '@code-dot-org/johnny-five';
-import {makeStubBoard, MicrobitStubBoard} from './makeStubBoard';
+import {makeStubBoard, MicrobitStubBoard} from './boards/makeStubBoard';
 import {
   PlaygroundButton,
   MicroBitButton
