+++ apps/test/unit/lib/kits/maker/LedTest.js
@@ -2,7 +2,7 @@
 import {expect} from '../../../../util/deprecatedChai';
 import sinon from 'sinon';
 import five from '@code-dot-org/johnny-five';
-import {makeStubBoard} from './makeStubBoard';
+import {makeStubBoard} from './boards/makeStubBoard';
 import Led from '@cdo/apps/lib/kits/maker/boards/circuitPlayground/Led';
 
 describe('Led', function() {
