+++ apps/test/unit/lib/kits/maker/boards/microBit/MicroBitBoardTest.js
@@ -1,8 +1,8 @@
-import {expect} from '../../../../util/deprecatedChai';
+import {expect} from '../../../../../../util/deprecatedChai';
 import MicroBitBoard from '@cdo/apps/lib/kits/maker/boards/microBit/MicroBitBoard';
-import {MicrobitStubBoard} from './makeStubBoard';
+import {MicrobitStubBoard} from '../makeStubBoard';
 import sinon from 'sinon';
-import {itImplementsTheMakerBoardInterface} from './MakerBoardTest';
+import {itImplementsTheMakerBoardInterface} from '../MakerBoardTest';
 import _ from 'lodash';
 import {EXTERNAL_PINS} from '@cdo/apps/lib/kits/maker/boards/microBit/MicroBitConstants';
 import ExternalLed from '@cdo/apps/lib/kits/maker/boards/microBit/ExternalLed';
